move functionality of "wego add" to separate package for coverage
